+++ src/validate/validate_projection.test.ts
@@ -19,20 +19,39 @@ describe('Validate projection', () => {
     test('Should return error in case of unknown property', () => {
         const errors = validateProjection({validateSpec: validate, value: {a: 1} as any, styleSpec: v8, style: {} as any});
         expect(errors).toHaveLength(1);
-        expect(errors[0].message).toContain('a');
-        expect(errors[0].message).toContain('unknown');
+        expect(errors[0].message).toContain('a: unknown property \"a\"');
     });
 
     test('Should return errors according to spec violations', () => {
         const errors = validateProjection({validateSpec: validate, value: {type: 1 as any}, styleSpec: v8, style: {} as any});
         expect(errors).toHaveLength(1);
-        expect(errors[0].message).toBe('type: expected one of [mercator, globe, vertical-perspective], 1 found');
+        expect(errors[0].message).toBe('projection: object expected, number found');
     });
 
-    test('Should pass if everything is according to spec', () => {
-        let errors = validateProjection({validateSpec: validate, value: {type: 'globe'}, styleSpec: v8, style: {} as any});
-        expect(errors).toHaveLength(0);
-        errors = validateProjection({validateSpec: validate, value: {type: 'mercator'}, styleSpec: v8, style: {} as any});
+    test('Should return error when value is null', () => {
+        const errors = validateProjection({validateSpec: validate, value: null as any, styleSpec: v8, style: {} as any});
+        expect(errors).toHaveLength(1);
+        expect(errors[0].message).toContain('projection: object expected, null found');
+    });
+
+    test('Should pass step function', () => {
+        const errors = validateProjection({validateSpec: validate, value: {'type': ['step', ['zoom'], 'vertical-perspective', 10, 'mercator']}, styleSpec: v8, style: {} as any});
         expect(errors).toHaveLength(0);
     });
+
+    test('Should pass string value', () => {
+        const errors = validateProjection({validateSpec: validate, value: {'type': 'mercator'}, styleSpec: v8, style: {} as any});
+        expect(errors).toHaveLength(0);
+    });
+
+    test('Should pass if [proj, proj, number]', () => {
+        const errors = validateProjection({validateSpec: validate, value: {'type': ['mercator', 'mercator', 0.3]}, styleSpec: v8, style: {} as any});
+        expect(errors).toHaveLength(0);
+    });
+
+    test('should parse interpolate', () => {
+        const errors = validateProjection({validateSpec: validate, value: {'type': ['interpolate', ['linear'], ['zoom'], 0, 'mercator', 5, 'vertical-perspective']}, styleSpec: v8, style: {} as any});
+        expect(errors).toHaveLength(0);
+    });
+
 });